@@ -1,8 +1,8 @@
 package dk.statsbiblioteket.larm_doms_exporter.util;
 
-import org.junit.Test;
-
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 public class FileResolverImplTest {
 
